Move DynamicCharAtlas to browser
